Found a missing database relation
The message I got:
ERROR:  relation "system_config" does not exist at character 128

I think immich-server runs some form of migrations that adds this to the
database. At the moment there is a chance that microservices starts
before the server. I hope this fixes that. It's still possible that I
need to add a small sleep to the microservices startup script. Time will
tell.

This is mostly a cosmetic error, the problem fixes it self.