Upgrade ffmpeg from 4.2.7 to 6.0
The snap is based on core20 (Ubuntu 20.04 LTS), that ships a older 4.x
release of ffmpeg. Immich has started to use newer features of ffmpeg so
I need to upgrade ffmpeg. This software has a lot of requirements and it
takes some work to build it, so I use pre-packaged binaries linked from
ffmpeg's official site.