WIP: new cli
This commit replaces the legacy old cli, with the new one included under
the cli folder, this is the recommended way to upload files.

I added the cli as part of the snap as an easy way upload files from the
sync feature, I exposed it as immich-distribution.cli because, why not,
but it has not been that useful because the sandbox limits what it can
read.

I will throw away the cli from this point on forward, and replace it with
a pure python implementation to upload files directly instead, this will
not be exposed to users.

I will commit this as a WIP commit if I choose to implement the cli
again in the future.