refactor, add node-gyp, fix sharp
This commits only new feature is that I have manually added node-gyp
before I install the server, so sharp builds against the locally
installed libvips, instead of using it's prebuild shipped binaries that
lack heic support.

As part of the troubleshooting I moved things around and moved the
server install code back to the snapcraft.yaml file, I also split the
server and web parts to separate parts. Custom patches and scripts that
I have written are separate parts as well.

Overall, I hope this is easier to maintain for the future. I also added
a check in the server build stage to make sure that we have a local
sharp binary.