Fix unstable verify_people_detected test
My guess is that there is some background process that need to be
finished for the API to return the correct number. I have added the
wait_for_empty_job_queue() function to make sure no jobs are executed in
the background, and then 10 retries before I finally fail the test.